Place Order Function Completed

+++ AAD_SPRING/12_Model_Mapper/src/main/java/lk/ijse/spring/service/impl/PurchaseOrderServiceImpl.java
@@ -1,8 +1,10 @@
 package lk.ijse.spring.service.impl;
 
-import lk.ijse.spring.dto.OrderDetailsDTO;
 import lk.ijse.spring.dto.OrdersDTO;
-import lk.ijse.spring.entity.*;
+import lk.ijse.spring.entity.Item;
+import lk.ijse.spring.entity.OrderDetails;
+import lk.ijse.spring.entity.OrderItem_PK;
+import lk.ijse.spring.entity.Orders;
 import lk.ijse.spring.repo.ItemRepo;
 import lk.ijse.spring.repo.OrderDetailsRepo;
 import lk.ijse.spring.repo.OrdersRepo;
@@ -14,7 +16,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
 /**
  * @author : Sanu Vithanage
@@ -38,52 +39,79 @@ public class PurchaseOrderServiceImpl implements PurchaseOrderService {
 
     @Override
     public void purchaseOrder(OrdersDTO dto) {
+        Orders order = mapper.map(dto, Orders.class);
         if (!ordersRepo.existsById(dto.getOid())) {
-           ordersRepo.save(new Orders(dto.getOid(),dto.getDate(),mapper.map(dto.getCustomer(), Customer.class)));
-            for (OrderDetailsDTO od : dto.getOrderDetails()) {
-                if (!orderDetailsRepo.existsById(new OrderItem_PK(od.getOid(),od.getItemCode()))) {
-                    orderDetailsRepo.save(new OrderDetails(od.getOid(), od.getItemCode(), od.getQty(), od.getUnitPrice()));
+            ordersRepo.save(order);
 
-                    Optional<Item> item = itemRepo.findById(od.getItemCode());
-                    if (item.isPresent()) {
-                        Item itm = item.get();
-                        itm.setQtyOnHand(itm.getQtyOnHand()-od.getQty());
-                        itemRepo.save(itm);
-                    }else{
-                        throw new RuntimeException("Transaction Rolled Back due to the Stock Update issue.!");
-                    }
+            if (dto.getOrderDetails().size() < 1) throw new RuntimeException("No items added for the order..!");
 
-                }else{
-                    throw new RuntimeException("Transaction Rolled Back due to the "+od.getItemCode()+" issue.!");
-                }
+            //update the item
+            for (OrderDetails orderDetail : order.getOrderDetails()) {
+                Item item = itemRepo.findById(orderDetail.getItemCode()).get();
+                item.setQtyOnHand(item.getQtyOnHand() - orderDetail.getQty());
+                itemRepo.save(item);
             }
-        }else{
-            throw new RuntimeException("Purchase Order Failed..!, Order ID "+dto.getOid()+" Already Exist.!");
+
+        } else {
+            throw new RuntimeException("Purchase Order Failed..!, Order ID " + dto.getOid() + " Already Exist.!");
         }
     }
 
     @Override
     public void deleteOrder(String oid) {
-    ordersRepo.deleteById(oid);
+        if (ordersRepo.existsById(oid)) {
+            ordersRepo.deleteById(oid);
+        } else {
+            throw new RuntimeException("Delete Order Failed..!, Order ID " + oid + " Not Exist..!");
+        }
+
     }
 
     @Override
     public void updateOrder(OrdersDTO dto) {
         if (ordersRepo.existsById(dto.getOid())) {
-            ordersRepo.save( mapper.map(dto,Orders.class));
 
-        }else{
-            throw new RuntimeException("Update Order Failed..!, Order ID "+dto.getOid()+" Not Exist.!");
+            Orders order = mapper.map(dto, Orders.class);
+            if (dto.getOrderDetails().size() < 1) throw new RuntimeException("No items added for the order..!");
+
+            for (OrderDetails od : order.getOrderDetails()) {
+                Item item = itemRepo.findById(od.getItemCode()).get();
+                OrderDetails previous = orderDetailsRepo.findById(new OrderItem_PK(od.getOid(), od.getItemCode())).get();
+
+                //Update the Item Qty
+                int newQty = od.getQty();
+                int prevQty = previous.getQty();
+                if (newQty > prevQty) {
+                    int dif = newQty - prevQty;
+                    item.setQtyOnHand(item.getQtyOnHand() - dif);
+                } else if (newQty < prevQty) {
+                    int dif = prevQty - newQty;
+                    item.setQtyOnHand(item.getQtyOnHand() + dif);
+                }
+                itemRepo.save(item);
+            }
+            //then delete the old order
+            ordersRepo.deleteById(dto.getOid());
+            //finally update the new order
+            ordersRepo.save(order);
+        } else {
+            throw new RuntimeException("Update Order Failed..!, Order ID " + dto.getOid() + " Not Exist.!");
         }
     }
 
     @Override
     public OrdersDTO searchOrder(String oid) {
-       return mapper.map(ordersRepo.findById(oid),OrdersDTO.class);
+        if (ordersRepo.existsById(oid)) {
+            return mapper.map(ordersRepo.findById(oid), OrdersDTO.class);
+        } else {
+            throw new RuntimeException("Search Order Failed..!, Order ID " + oid + " Not Exist.!");
+        }
+
     }
 
     @Override
     public List<OrdersDTO> getAllOrders() {
-        return mapper.map(ordersRepo.findAll(),new TypeToken<List<OrdersDTO>>(){}.getType());
+        return mapper.map(ordersRepo.findAll(), new TypeToken<List<OrdersDTO>>() {
+        }.getType());
     }
 }
